zm: Support mutable method arguments with dbus_interface macro
Compiling such an interface method resulted in a "expected expression"
error on the "mut" keyword. dbus_proxy did not have this issue, because
it only kept the argument name and not the whole pattern on the left of
the colon. So this commit changes dbus_interface to behave similarly.

Fixes #301