zb: Fix NoneBefore returns in MessageStream
Because MessageStream no longer receives updates on every message, its
internal tracking of next_seq is no longer valid.  However, if the
streams being joined originate from the same Connection (which is the
only reasonable way to use it), then any stream fed directly by the
sending task can answer NoneBefore if it is observed to be empty when
presented with another stream's MessageSequence.