zd: Refactor the attribute parsing logic.
Replace the getter style API for attribute parsing with a macro that generates separate structures for field and structure attributes that can be parsed once. This also adds checks for duplicate attributes and more clear error messages when an attribute is encountered in an unsupported location (e.g. a structure attribute on a field). Adding new attributes is as easy as writing 1-2 lines of code now.