zb: Drop internal use of ordered-stream
We started to make use of this (and infact ordered-stream was created
for zbus) when our API had user callbacks, which would mean messages on
the streams coming from multiple sources running in different
threads/tasks and hence we could not be sure that they'll always come in
the right order.

However, we dropped the callbacks long time ago and since then the
messages on all streams are pushed by the same task/thread: the socket
reader. So we do no need to worry about messages coming out of order.

Moreover, it turns out that [the "issue" I fixed in ordered-stream][1] to
make the new match-rule-based streams to work is actually an intentional
feature and the actual fix in zbus would [require some ugly code][2].
Given that we don't really need ordered-stream anymore, let's just not
rely on the ordered-stream internally anymore.

[1]: https://github.com/danieldg/ordered-stream/issues/12
[2]: https://gitlab.freedesktop.org/dbus/zbus/-/merge_requests/650