🚑️ zv: Drop 64-bit offsets support on 32-bit systems
If the native pointer size is 32-bit, then the 64-bit offset support is
not needed.

This also fixes the assumption that when a 32-bit offset is bumped, we
end up with a 64-bit offset. This is not true on 32-bit systems because
we use `usize` for offsets, which is 32-bit on 32-bit systems.

Fixes #357.