zb: add arg0ns to match rule
arg0namespace has an overly restrictive type (InterfaceName), which
disallows valid values (e.g. those containing '-' or single-element
names such as 'org').

Further more, a MatchRule with arg0namespace previously would not match
a message whose arg0 is not a valid InterfaceName (even if it is a valid
BusName).

fixes #382