🐛 zb: fdo::Monitoring::become_monitor to use MatchRule type
At the momen the method's arguments don't even have a good name and it's
not clear what they are. Let's fix this and use `MatchRule` type for the
match rules, to ensure users have a way to create rules in a safe way.

However, this is a breaking change. Doubtful that it will affect more
than 2 people on the planet at best though so maybe we just make the
call and fix it.