➖ zvariant: impl. `OwnedFd` and `RawFd` w/o libc
Zvariant implements a wrapper around the bare file descriptor types, so
`serde::Serialize` and `serde::Deserialize` can be implemented for it.

Before this change, `libc` was used to `clone` (`dup`), and `drop`
(`close`) an `OwnedFd`. The built-in `std::os::unix::io::OwnedFd` is
only stable since rust `1.63.0`, but by now the project's MSRV is
`1.64.0`, so we can use it.