🚑️ zb: Fix disconnect between cached & serialized serial number
OnceCell was being used to ensure that the serial number was only set
once but it did not take into account that a Message alwasy keeps its
serialized bytes and primary header must have a serial number set. So
until the serial number was set (either by user either sending it off
over the connection or calling `Connection::assign_serial_number`) the
serialized bytes will end up with a 0 serial number, while the message
would think it was never set.

This commit fixes this by making serial number a simple u32 everywhere
and treats the default value of 0 as "unset".