✨ zb: Add callback for socket closed
For the use case of a peer-to-peer D-Bus server, it's convenient to keep
a connection alive as long as the peer is connected.  This feature
allows a clean-up callback to automatically trigger the release of
server-side resources (including dropping the last Connection reference)
when the socket is closed by the remote client.