💥 zb: Split Socket trait into Read and Write halves
Socket implementation is now a bit more involved and require implementers
to split the type into read and write halves. This allows us to be able to
read and write to the same socket simultaneously. While that's more
efficient since read and write operations are not blocking each other,
this change will also allow us to convert raw::Connection and Socket API
to use async/await in the following commits.

Since we only have compile-only tests for vsock, implementations for
them has not actually been tested.

This is an API breaking change but hardly anyone will be affected by it.