💥 zv: Drop FD-specific serializer functions
Instead return

* the new `Encoded` type (which contains FDs on unix platforms) from
  to_bytes*.
* the FDs vector from to_writer*.

This also means that we switch from RawFd to OwnedFd type.

Caveat:

We have to dup the file descriptors but since Serialize::serialize takes
a reference to self, there is no way to get ownership right w/o dup.
However, dup is cheap so it's not a big deal IMO. Having said that, I do
have an idea on how to avoid the dup, but mostly likely it will require
additiion of a lifetime parameter to the `Message` type.