🥅 zv,zb: Value::to_owned now fallible as Value::try_to_owned
This means that a lot of conversions from Value and custom types are now
also fallible (i-e instead of `From` we have and require `TryFrom`).

We're doing this because w/o this change, when we switch to safe FD APIs,
we'll have no means to convert `Value::Fd` to an owned version. This is
because we'll need to use BorrowedFd::try_clone_to_owned, which is
fallible as dup() can fail easily and we need to let the user handle
this failure.