⬆️  Bump MSRV to 1.75 (the latest)
Our current MSRV (1.67) seems to have a bug:

https://github.com/dbus2/zbus/actions/runs/7379246404/job/20075278359?pr=533#step:5:91

```
error[E0700]: hidden type for `impl futures_core::Future<Output = std::result::Result<(), error::Error>>` captures lifetime that does not appear in bounds
   --> zbus/src/object_server/mod.rs:552:21
    |
552 |       ) -> Result<()> {
    |  _____________________^
553 | |         let root = self.root().read().await;
554 | |         let (node, manager_path) = root.get_child(&path, true);
555 | |         let node = node.unwrap();
...   |
587 | |         Ok(())
588 | |     }
    | |_____^
    |
    = note: hidden type `[async fn body@zbus/src/object_server/mod.rs:552:21: 588:6]` captures lifetime '_#22r
```

I could try and find the oldest version after 1.67 that works, but since
this is for our breaking release, I think we might as well bump to the
latest.

A possible positive side-effect might be that we may be able to shed the
`async-trait` dependency w/o breaking API (after 4.0 is out).