💚 zv: Higher recursion limit for test builds
We've a testcase that checks if ser/de of containers exceeding the
recursion limit, error out. Since these limits may not match that of the
default rustc limits, we can end up with rustc bailing on us[1].

Let's explicitly set a high enough Rust recursion limit for zvariant crate
when we're building for tests.

[1]: https://github.com/dbus2/zbus/actions/runs/7837026424/job/21385909120?pr=576