🚸 zb: Add connection::socket::WriteHalf::send_message method
This method is has a default implementation that just uses the
lower-level sibling method, `sendmsg`. Implementers can either override
this method or the lower-level method, depending on the underlying
socket they're using.

This will enable us to implement in-memory Connection pair (#591) and
hopefully be useful for other use cases too.