⚡️ zb: Pipeline client-side handshake as much as possible
We now send `DATA`, `NEGOTIATE_UNIX_FD` and `BEGIN` commands at once. This
reduces latency and hence improves performance.

It would have been the best to send all the commands at the same time,
but seems dbus-daemon kick us out if we send multiple commands after it
rejects the auth command. It wants us to negotiate the auth first. We
could in theory just reconnect but that would require some achitectural
changes and rethinking the code.

Moreover cookie auth requires a round-trip to the server to get
the cookie context first so we can't fully pipeline that one.

Furthermore, gdbus is not able to handle `BEGIN` sent along with `DATA`,
so we have to send `BEGIN` separately when gdbus feature is enabled on
Windows.

The server side is not possible to pipeline as it the server just
responds to commands from the client and it can not assume that clients
will always pipeline all commands.

Fixes #493.