🚑️ zx: Always use owned types for signal args
We started doing this already in cae1a65447 but it didn't cover all
cases and we still sometimes end up with types that do not implement
`serde::Deserialize`.

As a side-effect, this now also turns some of the normal method args
to take references instead of values, which is a good thing.

Fixes #412.