✨ zb: Add logging macros
I added several macros related to logging. This is intended to replace
the direct usage of tracing macros (which is what my pull request
addresses). These do not support all of the syntax of the standard
format_args macro, but I feel that they should still be comfortable to
work with, while remaining easy to understand and maintain. I have
inquired about the usage of info! because I personally prefer not to have
libraries using this more "general" log level directly, and I will remove
the info macro if I am allowed to do so.