🎨 zb: Implement log macros & cfg_attr instruments
This is a very large commit that touches many files in the zbus
codebase. However, it is minimally invasive. There are many clippy
warnings about unused variables, which will be addressed next. This
commit also changes the logging level of the instrumented functions and
futures to 'trace', notably excluding test functions. I added a custom
debug macro in e2e.rs because the zbus logging macros are not public.