🚨 zb: Feed unused logging variables to underscore
In CONTRIBUTING.md, it is stated that clippy must not find anything
wrong in the code, and that contributors may allow clippy to ignore
warnings. In a discussion I had with zeenix in pull request #712, we
discussed possible ways to allow unused variables. Feeding these
variables to an underscore ident inside of a block is a cheap way to
use these variables without forcing the user of the macros to worry
about unused variable clippy lints. This does require people to pass in
arguments themselves for the variables to be properly used.