🚨 zb: Underscore_prefix vars only used for logging
In CONTRIBUTING.md, it is stated that clippy must not find anything
wrong in the code, and that contributors may allow clippy to ignore
warnings. In a discussion I had with zeenix in pull request #712, we
discussed possible ways to allow unused variables. I decided that instead
of littering the codebase with the `#[allow(unused_variables)]`
annotation, I would instead prefix variables that have to do with logging
with an underscore. I have some misgivings about this, namely that it
breaks clippy's lint, but it is more maintainable this way. In
conclusion, I have no idea if this even *should* be allowed, but it gets
the job done and I'm not complaining!