🚩 zb: Support autolaunch address on Windows w/o `windows-gdbus` feature
Turns out that the reference dbus implementation supports autolaunch
fine and we don't need gdbus binary or special support for it.
Hopefully, in the near future, busd will also support it.

This commit drops the `windows-gdbus` feature. This means we're no longer
compatible with gdbus binary but that's not our fault that it can't handle
handshake pipelining. I seriously doubt anyone uses this anyway.

Fixes #687.