🎨 zb: Ensure all messsage bytes received during handshake are used
Don't assume that only a single full message could be received at the
end of the client handshake process. We may receive other messages too
if the server is very quick in sending them (or batches them). Worst
case scenario, we could also end up receiving partial messages, which
would currently completely break the connection state.

Let's be a bit smart about this and ensure that we consume all the bytes
that are received during the handshake process.

We are changing the type of the `already_received_bytes` arg of
`connection::socket::ReadHalf::receive_message` and theoretically, this
is an API break but given this is very new API, it's very unlikely that
anyone implements Socket anyway and this change is needed to fix the
issue at hand, let's not worry about it.