✨ zv: support ser/de of values directly
This change is inspired by similiar functionality in serde-json
which allows for types to be serialized/deserialized into an
AST directly. This is useful for a variety of reasons, and I
think would be of benefit to zbus as well.

Luckily `zvariant` already has a type pretty close to this -
the `zvariant::Value`. This thing can represent any value that
dbus can represent on the wire. It can also model any variant
received from some other service. By serializing and deserializing
directly to this thing - it gives clients stronger capability
to deal with variants they receive.