🔨 add githook scripts
This change adds githook scripts to the `.githooks` folder so they
can be more trivially enabled by developers working on this repo
by using the command `git --local core.hooksPath .githooks/`. They
do the same basic function as the scripts documented in
`Contributing.md` but with a little extra flair.