zv: migrate dbus ser to using ParsedSignature
This change moves the DBus Serializer implementation to using the
new signature parser. This removes from the serializer code the
extra signature parsing that was happening during the process
of deserialization - leading to big simplifications of the code.

In addition, it should be easier to handle new use cases by
pattern matching on the signature as needed.