🔥 zb: Drop use of macro_rules for fdo proxies
Using macro_rules to then generate the actual type through a proc macro
is a bit strange anyway and might be giving us problems with
rust-analyzer (#671).

Instead just generate both the async and blocking proxies from the same
call and then just alias the blocking one from the `blocking::fdo` mod.

This does mean that we now have redundant blocking proxy types in `fdo`.
This will be fixed in a following commit.