✨ zv: Add parsed::Signature
This is a new type that represents a parsed signature. We'll use this to
avoid parsing the signature as part of (de)serialization, which seems to
be quite expensive at times, especially for large arrays.

While the the parsed::Signature can be constructed w/o any allocation
and in const context even, unfortunately converting from and to
`Signature` or the underlying string form, is not free and will
require allocations for non-basic types.

However, we'll be changing `Type` trait to provide us a
`parsed::Signature` as a constant, allowing us to avoid allocations and
parsing.