⚡️ zv,zd,zn,zb,zm: Type::parsed_signature now a const
Convert `Type::parsed_signature` to a const, named `SIGNATURE`. We also
remove `Type::signature` method, since `parsed::Signature` can be easily
converted to a `crate::Signature`. This commit makes similar changes to
`DynamicType` and `DynamicDeserialize` as well.

This is an API break that breaks all code that implements `Type`
(manually), `DynamicType` and/or `DynamicDeserialize`. In return, we
eliminate a lot of allocations, which can make a big difference when it
comes to (de)serialization on low-end machines.

Fixes #882.