✨ zv,zd,zn,zb,zm: Add Type::parsed_signature()
Add a method to `Type` trait that returns the parsed form of the
signature. This method has a default implementation to not break the API.
By default, it parses the return value of the `signature()` method. Since
that is not efficient at all, we encourage implementors to implement this
method (instead of `signature()`). Moreover, for keeping type
implementation easy, we also now provide a default implementation of
`signature`, which converts the parsed signature back to a string form.

This does indeed create a mutual recursion between the default
implementations of `signature` and `parsed_signature`, but I believe this
is a good trade-off for the sake of simplicity and ease of implementation.
All current manul implementations of `Type` out there are guaranteed to
have an implementation of `signature` method, so this will only affect new
implementations. Hopefully they will read the docs when they implement the
trait. If not, they'll quickly run into a stack overflow, which will force
them to read the docs. :)

This commit also replaces most implementations of `signature` with
`parsed_signature`. Only manual implementations of `Type` outside this
repository will want to (not need to) implement `parsed_signature` if they
want encoding/decoding to be super efficient efficient.

I originally wrote this as a const (as it should be) but that will need to
have a default value to not break the API and from a consts context, we
can't call `signature`, even if we could parse its return value. This will
be an easy fix to add in the next major version.