🔥 zv: Drop all use of internal SignatureParser
Instead make use of the new `parsed::Signature` API.

This involves breaking the API a bit since now both `ArraySeed` and
`StructureSeed` do not implement `Sync` and `Send` now. However the
chances of anyong using them directly are already extremely low and even
if they'd use them directly, the changes of them needing `Sync` and
`Send` are even lower. So I think we'll be ok.