⚡️ zv: Avoid an allocation in signature parser
Avoid the allocation of `Vec<Signature>` in case of a single signature on
the top-level. This is a a very common case, especially in variants, where
the signature needs to be parsed at runtime.

On my beefy machine the performance gains were not huge (around 5-8%) and
not very consistent but I suspect they'll be more noticeable on low-end
machines.