⚡️ zv: Use `Box` instead of `Rc` in Signature
While this may seem like this would make the performance worse, it
actually makes it better in some of the benchmarks, while most of the
benbenchmarks show no noticeable difference with this change.

Moeover, this make the Signature type Send + Sync, which is something we
are going to need to keep a Signature in public types that implement Send
and Sync already and we want to keep them that way.