✨ zd: Validate user provided signature
The ability to just specify the signature string as an attribute through
`Type` derive macro is very convenient but weird things can happen at
runtime if the user makes a mistake.

Now that zvariant_utils provide the API to validate the signature, let's
make use of that to ensure user didn't make a mistake at compile-time.