💥 zm,zb: `proxy` macro respects visibility
This includes all types generated by `proxy`.

Unfortunately this means that the existing code will have to set the
visiblity explicitly to `pub` if they were relying on the generated
proxy to be public. Good that we're in API break.

For now, the `interface` hardcodes visibility of the generated proxy (if
any) to `pub` until we add a way to control it in `interface` macro.