🚑️ zv: More lenient check for signature match
In `DynamicDeserialize` impl for `Type`, if the given signature doesn't
match the expected signature and the expected signature is a struct with
a single field, match against the field's signature instead. This is
needed when D-Bus message body consists of a single type and it is being
deserialized as a single-field struct.

Fixes #1015.