✨ zm: interface now provides convenience API to emit signals
`interface` now generates a trait, <InterfaceType>Signals, that provides
the same signal methods as user specifies but w/o the `SignalEmitter`
argument. The macro also generates 2 implementations of this trait for:

* InterfaceRef<InterfaceType>
* SignalEmitter.

The former is useful for emitting signals from outside the context of an
interface method and the latter is useful for emitting signals from
inside an interface methods.

Fixes #871.