🛂 zb: Use string type for auth mechanisms in Rejected cmd
Now that we don't support all authentication mechanisms described in the
spec, we'll get a parse error on receiving an unsupported mechanism.
Let's use the raw string type for the mechanism field in the Rejected
command to avoid this.

We do so in a way that limits allocations, especially on the
server-side, which might need to reject a lot of connections and it will
have more impact there.