⏪️ Revert "Merge pull request #989 from elmarco/address"
This reverts commit fc03642b0387ef1e6e21326573ae364148460167, reversing
changes made to deffbb8061a87f4b1dfc7a22547ce55bdb2a2cc8.

It turns out there was a major design flaw in the API: the parsed
address string was not kept in a parsed state, resulting in reparsing of
the string form, for each getter. Because of this, I decided to revert
the changes for 5.0.