♻️  zv: Use winnow for validation of ObjectPath
This makes it so **much** simpler. It unfortunately also makes it a bit
slower (~6%) but we're talking a few nanoseconds here so it will not make
a huge difference overall. For example, the zbus benchmarks show no
difference at all, even header parsing, which does contain object paths.

There is also a good chance that the performance will improve in the next
version of winnow. See the dicsussion here:

https://github.com/winnow-rs/winnow/discussions/610

Partially fixes #1070.