♻️  zv: Use winnow for validation of unique names
Just like in the case for object paths in a previous commit in this
series, this change makes the parsing code **much** simpler and more
readable. However, unlike in case of the object path, it seems to make
parsing **much** faster (~40%). Unlike the case for the similar change
for well-known name parsing, the header parsing benchmark shows a 6%
improvement in performance. This is extremely strange, since unique
names are not being used in the benchmark. However, given that this
balances out the performance regression from the previous change and that
we're still talking a difference of around 10 nanoseconds here, I don't
think there is anything to worry about.

A "regression" here is the error message: it's now very generic. However,
it links to the relavent section in the specification so people will have
a way to figure it out.

Partially fixes #1070.