👷 CI: Check doc build for individual packages separately
Currently we end up with a strange error of enabling all features in the
root workspace. I don't know what really causes it (could be a Cargo
bug) but I also don't see the need to enable all features in the root so
I don't want to bother with that.