🚸 zn: Use Error::InvalidName variant for invalid name everywhere
Our users don't need to know exactly what type of the name couldn't be
constructed. Typically they'll get these errors when they are explicitly
converting from a string to a specific name type and will hence know
which name type they were trying to construct.

The contained string provides information on the exact type of name that
failed to be constructed anyway.