👷 CI: Run benchmarks as part of the CI on pushes to main
So if a change brings any performane regressions, we can catch them early on.

I originally wanted to run this workflow on PRs but seems that's not
only not possible from forked repos[1] but also a security hole[2]. Oh
well, at least we can catch the regressions on the main repo.

[1]: https://github.com/benchmark-action/github-action-benchmark/issues/56
[2]: https://github.com/benchmark-action/github-action-benchmark?tab=readme-ov-file#run-only-on-your-branches