➖ zb: Tie async-process dep to async-io feature
instead of the target OS. Originally, this dep was only needed by Mac,
which is not our prime target so it was ok to tie its use to Mac. But
now we need it also on Linux (our prime target) but only if `async-io`
is needed. This means that with the current setup, we end up making
tokio users (a big majority of our users) also needlessly depend on it.