⚡️ zm: interface allows to disable docs in introspection
Introspection strings are part of the binary and can have an impact on
the binary size. This change allows to disable the addition of docs in the
introspection for cases where there interface is well-known/well-documented.

For example, combined with the following change of using this in the
`zbus::fdo` interfaces, this will allow `busd` binary to shed 26 KB.