⚡️ zb: Make header cloning cheaper
Right now cloning a header involves cloning the `PrimaryHeader` (cheap,
mostly POD), and then cloning the `Fields`, which can be somewhat
complex, as it involves cloning all the strings and the signature, which
if dynamic could be rather costly.

From things coming from a message, strings are always `Borrowed` so the
cloning is cheap, so the only expensive thing is cloning the signature.

Using `Cow` for it looks sensible, it's what `Str` is doing internally
already, effectively.