🐛 zv: fix wrong lifetime when creating a Value from &Vec and &[]
When creating a Value from a borrowed vec or slice using Value's From
implementation, the compiler would wrongfully assume, that the created
value has the same lifetime, even though the input gets cloned when
creating the Array struct. This commit makes the impl assign
a new lifetime to the cloned value.