✨ zb,zn,zv: Implement Borrow for Owned* types
Implementing `Borrow` makes things such as using `Owned*` types as
keys easier, and it was more-or-less designed for this type of scenario
anyway. Although it comes with some stringent requirements (i.e. traits
implemented on the owned type need to have identical behavior to the
borrowed type), all of these types are already rather thin wrappers
to begin with.