🚑️ zv: More robust deserialization of variants
Our deserializers rely on a specific behavior from the Value's
Deserialize impls, and that fails when Value is being deserialized in a
generic way (through deserialize_any or deserialize_ignored_any) and we
end up with overflowing the stack.

This happens for example when ignoring unknown fields/entries.

Unfortunately, in case of GVariant we now no longer error out if a null
byte is missing in a string type. You can blame Allison Lortie for
removing the null-termination from a signature **only** when it is encoded
in a variant but not otherwise. :)