🧵 zb: Remove deadlocks in Connection name request tasks
We were using a `MessageStream` to receive `NameAcquired` and `NameLost`
signals in tasks that were kept the in the `Connection`'s inner type and
that meant we'd end up leaking `Connection` instance in certain cases.

With the `RequestNameFlags::AllowReplacement` becoming a default flag,
this issue is now even 100% reproducible in typical scenarios.

Let's use the internal API that avoid creating any cyclic references and
fixes the problem. This is the same solution we employ for the
on-demand `ObjectServer` task.