✨ zb: Add an interface_unchecked function to object_server
The current interface function requires a read lock to be acquired on the interface before returning the references. It needs this to check if the interface can be safely downcasted. This new function does not contain that check allowing it be used both internally and externally, to acquire a shared reference to the current interface. This is especially useful when annotating the functions with signalreceiver is not possible or desired